+++ src/main/java/nl/novi/catsittermanager/services/CatSitterService.java
@@ -28,25 +28,4 @@ public class CatSitterService {
         this.orderRepository = orderRepository;
         this.orderService = orderService;
     }
-
-    public CatSitterDto transferToDto(CatSitter catSitter) {
-
-        CatSitterDto catSitterDto = new CatSitterDto();
-
-        // catSitterDto.id = catSitter.getId(); // deze komt uit superklasse User
-        catSitterDto.about = catSitter.getAbout();
-        catSitterDto.customers = catSitter.getCustomers();
-        catSitterDto.orderList = catSitter.getOrderList();
-
-        return catSitterDto;
-    }
-
-    public CatSitter transferToCatSitter(CatSitterInputDto catSitterDto) {
-
-        CatSitter catSitter = new CatSitter();
-
-        catSitter.setAbout(catSitterDto.about);
-        catSitter.setCustomers(catSitterDto.customers);
-        catSitter.setOrderList(catSitterDto.orderList);
-    }
 }
