+++ src/main/java/nl/novi/catsittermanager/services/CustomerService.java
@@ -36,28 +36,4 @@ public class CustomerService {
         this.customerService = customerService;
     }
 
-    public CustomerDto transferToDto(Customer customer) {
-
-        CustomerDto customerDto = new CustomerDto();
-
-        // catSitterDto.id = catSitter.getId(); // deze komt uit superklasse User
-        customerDto.numberOfCats = customer.getNumberOfCats();
-        customerDto.orderList = customer.getOrderList();
-        customerDto.catListByName = customer.getCatListByName();
-        customerDto.catsitters = customer.getCatSitters();
-
-        return customerDto;
-    }
-
-    public Customer trasnferToCustomer(CustomerInputDto customerDto) {
-
-        Customer customer = new Customer();
-
-        customer.setNumberOfCats(customerDto.numberOfCats);
-        customer.setOrderList(customerDto.orderList);
-        customer.setCatListByName(customerDto.catListByName);
-        customer.setCatSitters(customerDto.catsitters);
-
-        return customer;
-    }
 }
