+++ src/main/java/nl/novi/catsittermanager/services/InvoiceService.java
@@ -21,28 +21,4 @@ public class InvoiceService {
         this.orderRepos = orderRepos;
         this.orderService = orderService;
     }
-
-    public InvoiceDto TransferToDto(Invoice invoice) {
-
-        InvoiceDto invoiceDto = new InvoiceDto();
-
-        invoiceDto.invoiceNo = invoice.getInvoiceNo();
-        invoiceDto.invoiceDate = invoice.getInvoiceDate();
-        invoiceDto.amount = invoice.getAmount();
-        invoiceDto.order = invoice.getOrder();
-
-        return invoiceDto;
-    }
-
-    public Invoice TransferToInvoice(InvoiceInputDto invoiceDto) {
-
-        Invoice invoice = new Invoice();
-
-        invoice.setInvoiceNo(invoiceDto.invoiceNo);
-        invoice.setInvoiceDate(invoiceDto.invoiceDate);
-        invoice.setAmount(invoiceDto.amount);
-        invoice.setOrder(invoiceDto.order);
-
-        return invoice;
-    }
 }
