+++ src/main/java/nl/novi/catsittermanager/services/TaskService.java
@@ -22,31 +22,4 @@ public class TaskService {
         this.orderRepos = orderRepos;
         this.orderService = orderService;
     }
-
-    public TaskDto transferToDto(Task task) {
-
-        TaskDto taskDto = new TaskDto();
-
-        taskDto.id = task.getId();
-        taskDto.taskType = task.getTaskType();
-        taskDto.taskInstruction = task.getTaskInstruction();
-        taskDto.extraInstructions = task.getExtraInstructions();
-        taskDto.priceOfTask = task.getPriceOfTask();
-        taskDto.order = task.getOrder();
-
-        return taskDto;
-    }
-
-    public Task transferToTask(TaskInputDto taskDto) {
-
-        Task task = new Task();
-
-        task.setTaskType(taskDto.taskType);
-        task.setTaskInstruction(taskDto.taskInstruction);
-        task.setExtraInstructions(taskDto.extraInstructions);
-        task.setPriceOfTask(taskDto.priceOfTask);
-        task.setOrder(taskDto.order);
-
-        return task;
-    }
 }
